Refactor controllers for better code quality

+++ web/controllers/passwordresets_controller.ex
@@ -1,4 +1,6 @@
 defmodule Thegm.PasswordResetsController do
+  @moduledoc "Controller responsible for handling password resets"
+
   use Thegm.Web, :controller
 
   alias Thegm.PasswordResets
@@ -9,21 +11,19 @@ defmodule Thegm.PasswordResetsController do
       {"resets", params} ->
         user = Repo.get_by(Users, email: params["email"])
 
-        cond do
-          user ->
-            Repo.delete_all(PasswordResets, [user_id: user.id, used: false])
-            changeset = PasswordResets.changeset(%PasswordResets{},%{"used" => false, "user_id" => user.id})
+        if user do
+          Repo.delete_all(PasswordResets, [user_id: user.id, used: false])
+          changeset = PasswordResets.changeset(%PasswordResets{}, %{"used" => false, "user_id" => user.id})
 
-            case Repo.insert(changeset) do
-              {:ok, reset} ->
-                Thegm.Mailgun.email_password_reset(params["email"], reset.id)
-                |> Thegm.Mailer.deliver_now
-                send_resp(conn, :created, "")
-            end
+          case Repo.insert(changeset) do
+            {:ok, reset} ->
+              Thegm.Mailgun.email_password_reset(params["email"], reset.id)
+              |> Thegm.Mailer.deliver_now
+              send_resp(conn, :created, "")
+          end
 
-          true ->
-            send_resp(conn, :created, "")
-
+        else
+          send_resp(conn, :created, "")
         end
 
       _ ->
@@ -33,6 +33,7 @@ defmodule Thegm.PasswordResetsController do
     end
   end
 
+
   def update(conn, %{"id" => id, "data" => %{"attributes" => params}}) do
     case Repo.get(PasswordResets, id) do
       nil ->
@@ -40,38 +41,41 @@ defmodule Thegm.PasswordResetsController do
         |> put_status(:not_found)
         |> render(Thegm.ErrorView, "error.json", errors: ["Invalid password reset"])
       resp ->
-        cond do
-          resp.used == false && NaiveDateTime.diff(NaiveDateTime.utc_now(), resp.inserted_at) ->
-            code = PasswordResets.changeset(resp, %{used: true})
-            case Repo.update(code) do
-              {:ok, updated_code} ->
-                case Repo.get(Thegm.Users, updated_code.user_id) do
-                  nil ->
-                    conn
-                    |> put_status(:not_found)
-                    |> render(Thegm.ErrorView, "error.json", errors: ["Unable to locate user"])
-                  user ->
-                    user = Thegm.Users.changeset(user, %{password: params["password"]})
-                    case Repo.update(user) do
-                      {:ok, _} ->
-                        send_resp(conn, :ok, "")
-                      {:error, user} ->
-                        conn
-                        |> put_status(:bad_request)
-                        |> render(Thegm.ErrorView, "error.json", errors: Enum.map(user.errors, fn {k, v} -> Atom.to_string(k) <> ": " <> elem(v, 0) end))
-                    end
-                end
-              {:error, updated_code} ->
-                conn
-                |> put_status(:bad_request)
-                |> render(Thegm.ErrorView, "error.json", errors: Enum.map(updated_code.errors, fn {k, v} -> Atom.to_string(k) <> ": " <> elem(v, 0) end))
-            end
+        if resp.used == false && NaiveDateTime.diff(NaiveDateTime.utc_now(), resp.inserted_at) do
+          code = PasswordResets.changeset(resp, %{used: true})
+          case Repo.update(code) do
+            {:ok, updated_code} ->
+              new_password = params["password"]
+              update_password(conn, user, new_password)
+            {:error, updated_code} ->
+              conn
+              |> put_status(:bad_request)
+              |> render(Thegm.ErrorView, "error.json", errors: Enum.map(updated_code.errors, fn {k, v} -> Atom.to_string(k) <> ": " <> elem(v, 0) end))
+          end
 
-          true ->
+        else
+          conn
+          |> put_status(:not_found)
+          |> render(Thegm.ErrorView, "error.json", errors: ["Invalid password reset"])
+      end
+    end
+  end
+
+  def update_password(conn, user, password) do
+    case Repo.get(Thegm.Users, user.id) do
+      nil ->
+        conn
+        |> put_status(:not_found)
+        |> render(Thegm.ErrorView, "error.json", errors: ["Unable to locate user"])
+      user ->
+        user = Thegm.Users.changeset(user, %{password: password})
+        case Repo.update(user) do
+          {:ok, _} ->
+            send_resp(conn, :ok, "")
+          {:error, user} ->
             conn
-            |> put_status(:not_found)
-            |> render(Thegm.ErrorView, "error.json", errors: ["Invalid password reset"])
-
+            |> put_status(:bad_request)
+            |> render(Thegm.ErrorView, "error.json", errors: Enum.map(user.errors, fn {k, v} -> Atom.to_string(k) <> ": " <> elem(v, 0) end))
         end
     end
   end
